git/website: Add new update button to DownloadList component
This commit adds a new button called "New Update" to the DownloadList component in order to indicate when there is a new update available for a particular item. The button is displayed only if the "new-update"
property of the item is set to "yes". This enhancement improves the user experience by providing a clear visual indication of new updates.

Signed-off-by: KARTHIK LAL <karthik.lal558@gmail.com>