git/website: update "New update available" button component with expiryModule
The code changes add a new button called "New Update Available" to the DownloadList component. This button is displayed only if the "new-update" property of the item is
set to "yes" and the update is within the last 3 days. This enhancement improves the user experience by providing a clear visual indication of new updates.

Signed-off-by: KARTHIK LAL <karthik.lal558@gmail.com>